api_yamdb/requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.0.12
colorama==0.4.5
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.0.4
idna==3.4
iniconfig==1.1.1
packaging==21.3
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.11.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.2.1
requests==2.26.0
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.3
urllib3==1.26.12
